Scrub secure info from job logs
